Item handlings module, comparing product amount with real store by importing csv file features were  added successfully. Edit some bugs and issues on task timeline.

+++ application/controllers/Units.php
@@ -1,51 +1,53 @@
 <?php
-
 defined('BASEPATH') OR exit('No direct script access allowed');
 
-class Units extends CI_Controller{
+class Units extends CI_Controller
+{
 
-    public $viewFolder   = "";
-    public $pageTitle    = "";
+    public $viewFolder = "";
+    public $pageTitle = "";
     public $pageTitleExt = "";
 
     public function __construct()
     {
         parent::__construct();
-        $this->viewFolder   = "units_v";
-        $this->pageTitle    = "Ölçü Vahidləri";
+        if(!get_active_user()){
+            redirect(base_url("login"));
+        }
+        $this->viewFolder = "units_v";
+        $this->pageTitle = "Ölçü Vahidləri";
         $this->pageTitleExt = PageTitleExt;
-
         $this->load->model('units_model');
-
     }
 
     public function index(){
-
         $viewData = new stdClass();
-        $viewData->viewFolder    = $this->viewFolder;
-        $viewData->subViewFolder = "list";
-        $viewData->pageTitle     = $this->pageTitle.$this->pageTitleExt;
-        $viewData->items         = $this->units_model->getAll();
+        $viewData->viewFolder       = $this->viewFolder;
+        $viewData->subViewFolder    = "list";
+        $viewData->pageTitle        = $this->pageTitle.$this->pageTitleExt;
+        $viewData->header           = $this->pageTitle;
 
-        $this->load->view("{$viewData->viewFolder}/{$viewData->subViewFolder}/list/index",$viewData);
+        $this->load->view("{$viewData->viewFolder}/{$viewData->subViewFolder}/index",$viewData);
     }
 
     public function new_form(){
 
         $viewData = new stdClass();
-        $viewData->viewFolder    = $this->viewFolder;
-        $viewData->subViewFolder = "teachers";
-        $viewData->pageTitle     = "Müəllim Əlavə Et".$this->pageTitleExt;
+        $viewData->viewFolder       = $this->viewFolder;
+        $viewData->subViewFolder    = "add";
+        $viewData->pageTitle        = "Ölçü vahidi Əlavə Et".$this->pageTitleExt;
+        $viewData->header           = "Ölçü vahidi Əlavə Et";
 
-        $this->load->view("{$viewData->viewFolder}/{$viewData->subViewFolder}/add/index",$viewData);
-
+        $this->load->view("{$viewData->viewFolder}/{$viewData->subViewFolder}/index",$viewData);
     }
 
     public function save(){
 
         $this->load->library('form_validation');
 
-        $this->form_validation->set_rules("fullname","Müəllim Adı","required|trim");
+
+        $this->form_validation->set_rules("unit-name","Ölçü vahidi adı","required|trim");
+        $this->form_validation->set_rules("unit-short-name","Qısa adı","required|trim");
 
         $this->form_validation->set_message(array(
             "required" => "{field} boş buraxıla bilməz!"
@@ -54,11 +56,9 @@ class Units extends CI_Controller{
         $checkValidation = $this->form_validation->run();
 
         if($checkValidation){
-
             $save = $this->units_model->add(array(
-                "fullName"   => $this->input->post('fullname'),
-                "createdAt" => date("Y-m-d H:i:s"),
-                "createdBy" => $this->session->userdata('admin-login')->ID
+                "name"          => $this->input->post("unit-name"),
+                "shortName"     => $this->input->post("unit-short-name")
             ));
 
             if($save){
@@ -69,7 +69,6 @@ class Units extends CI_Controller{
                     "position" => "toast-top-center"
                 );
                 $this->session->set_flashdata("alert",$alert);
-                redirect(base_url('teachers'));
             }
             else{
 
@@ -80,49 +79,47 @@ class Units extends CI_Controller{
                     "position" => "toast-top-center"
                 );
                 $this->session->set_flashdata("alert",$alert);
-                redirect(base_url('teachers/add-teacher'));
             }
-
-
+            redirect(base_url('units/add-unit'));
         }
         else{
 
             $viewData = new stdClass();
-            $viewData->viewFolder    = $this->viewFolder;
-            $viewData->subViewFolder = "teachers";
-            $viewData->pageTitle     = "Müəllim Əlavə Et".$this->pageTitleExt;
-            $viewData->form_error    = true;
+            $viewData->viewFolder       = $this->viewFolder;
+            $viewData->subViewFolder    = "add";
+            $viewData->pageTitle        = "Ölçü vahidi Əlavə Et".$this->pageTitleExt;
+            $viewData->header           = "Ölçü vahidi Əlavə Et";
+            $viewData->form_error       = true;
 
-            $this->load->view("{$viewData->viewFolder}/{$viewData->subViewFolder}/add/index",$viewData);
+            $this->load->view("{$viewData->viewFolder}/{$viewData->subViewFolder}/index",$viewData);
         }
-
     }
 
     public function update_form($id){
 
         $viewData = new stdClass();
-        $viewData->viewFolder    = $this->viewFolder;
-        $viewData->subViewFolder = "teachers";
-        $viewData->pageTitle     = "Müəllim Redaktə Et".$this->pageTitleExt;
-
-        $getData = $this->units_model->get(array(
-            "ID" => $id
+        $viewData->viewFolder       = $this->viewFolder;
+        $viewData->subViewFolder    = "update";
+        $viewData->pageTitle        = "Ölçü vahidi Redaktə Et".$this->pageTitleExt;
+        $viewData->header           = "Ölçü vahidi Redaktə Et";
+        $viewData->unit             = $this->units_model->get(array(
+            "ID"=>$id
         ));
+        if(!$viewData->unit):
+            $viewData->result     = false;
+        else:
+            $viewData->result     = true;
+        endif;
 
-        if($getData){
-            $viewData->item = $getData;
-        }
-        else{
-            redirect(base_url('teachers'));
-        }
-
-        $this->load->view("{$viewData->viewFolder}/{$viewData->subViewFolder}/update/index",$viewData);
+        $this->load->view("{$viewData->viewFolder}/{$viewData->subViewFolder}/index",$viewData);
     }
 
     public function update($id){
+
         $this->load->library('form_validation');
 
-        $this->form_validation->set_rules("fullname","Müəllim Adı","required|trim");
+        $this->form_validation->set_rules("unit-name","Ölçü vahidi adı","required|trim");
+        $this->form_validation->set_rules("unit-short-name","Qısa adı","required|trim");
 
         $this->form_validation->set_message(array(
             "required" => "{field} boş buraxıla bilməz!"
@@ -131,16 +128,16 @@ class Units extends CI_Controller{
         $checkValidation = $this->form_validation->run();
 
         if($checkValidation){
-
-            $save = $this->units_model->update(
+            $update = $this->units_model->update(
                 array(
-                    "ID" => $id
+                  "ID" => $id
                 ),
                 array(
-                    "fullName"   => $this->input->post('fullname'),)
-            );
+                    "name"          => $this->input->post("unit-name"),
+                    "shortName"     => $this->input->post("unit-short-name")
+            ));
 
-            if($save){
+            if($update){
                 $alert = array(
                     "title"    => "Əməliyyat uğurla yerinə yetirildi",
                     "text"     => "",
@@ -148,7 +145,6 @@ class Units extends CI_Controller{
                     "position" => "toast-top-center"
                 );
                 $this->session->set_flashdata("alert",$alert);
-                redirect(base_url('teachers'));
             }
             else{
 
@@ -158,50 +154,67 @@ class Units extends CI_Controller{
                     "type"     => "error",
                     "position" => "toast-top-center"
                 );
+
                 $this->session->set_flashdata("alert",$alert);
-                redirect(base_url('teachers'));
             }
-
-
+            redirect(base_url('units'));
         }
         else{
 
             $viewData = new stdClass();
-            $viewData->viewFolder    = $this->viewFolder;
-            $viewData->subViewFolder = "teachers";
-            $viewData->pageTitle     = "Müəllim Redaktə Et".$this->pageTitleExt;
-            $viewData->form_error    = true;
-            $viewData->item          = $id;
+            $viewData->viewFolder       = $this->viewFolder;
+            $viewData->subViewFolder    = "update";
+            $viewData->pageTitle        = "Ölçü vahidi Əlavə Et".$this->pageTitleExt;
+            $viewData->header           = "Ölçü vahidi Əlavə Et";
+            $viewData->form_error       = true;
 
-            $this->load->view("{$viewData->viewFolder}/{$viewData->subViewFolder}/update/index",$viewData);
+            $this->load->view("{$viewData->viewFolder}/{$viewData->subViewFolder}/index",$viewData);
         }
     }
 
     public function delete($id){
 
-        $delete = $this->units_model->delete(array(
-            "ID" => $id
+        $deletable = $this->units_model->getAll(array(
+           "isDeletable" => 1
         ));
+        $deletableUnits = array();
+        foreach ($deletable as $del){
+            array_push($deletableUnits, $del->ID);
+        }
 
-        if($delete){
-            $alert = array(
-                "title"    => "Əməliyyat uğurla yerinə yetirildi",
-                "text"     => "Məlumat silindi",
-                "type"     => "success",
-                "position" => "toast-top-center"
-            );
+        if(in_array($id, $deletableUnits)){
+            $delete = $this->units_model->delete(array(
+                "ID" => $id
+            ));
+
+            if($delete){
+                $alert = array(
+                    "title"    => "Əməliyyat uğurla yerinə yetirildi",
+                    "text"     => "Məlumat silindi",
+                    "type"     => "success",
+                    "position" => "toast-top-center"
+                );
+            }
+            else{
+                $alert = array(
+                    "title"    => "Üzr istəyirik!",
+                    "text"     => "Bilinməyən xəta baş verdi. Zəhmət olmasa birazdan bir daha cəhd edin",
+                    "type"     => "error",
+                    "position" => "toast-top-center"
+                );
+            }
         }
         else{
             $alert = array(
                 "title"    => "Üzr istəyirik!",
-                "text"     => "Bilinməyən xəta baş verdi. Zəhmət olmasa birazdan bir daha cəhd edin",
+                "text"     => "Bu əməliyyat sistem tərəfindən qadağan edilib",
                 "type"     => "error",
                 "position" => "toast-top-center"
             );
         }
 
         $this->session->set_flashdata("alert",$alert);
-        redirect(base_url('teachers'));
+        redirect(base_url('units'));
     }
 
     public function isActiveSetter($id)
@@ -217,6 +230,10 @@ class Units extends CI_Controller{
                 )
             );
         }
+    }
+
+    public function getDataTable(){
+        echo $this->units_model->getDataTable();
     }
 
 }
